MCOL-5001 EM code preparation to merge EM and PP runtimes
